writeBufferToDisk

+++ internal/engine/core/core.go
@@ -24,7 +24,7 @@ var storageBlock sync.RWMutex
 
 type tCoreSettings struct {
 	Storage      string
-	BucketSize   int
+	BucketSize   int64
 	FriendlyMode bool
 }
 
@@ -155,7 +155,7 @@ type tColumnInfo struct {
 	Folder        string               `json:"folder"`
 	Parents       string               `json:"parents"`
 	BucketLog     uint8                `json:"blog"`
-	BucketSize    int                  `json:"bsize"`
+	BucketSize    int64                `json:"bsize"`
 	OldRev        string               `json:"oldrev"`
 	CurrentRev    string               `json:"currentrev"`
 	Specification TColumnSpecification `json:"specification"`
@@ -183,8 +183,8 @@ type TColumnForWrite struct {
 
 type tColumnForStore struct {
 	Field string
-	Id    uint64
-	Time  int64
+	Id    uint64 // FIXME: Need delete
+	Time  int64  // FIXME: Need delete
 	Value string
 }
